Move DXT decoding to GPU/Common/.
